+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 Versions
 ========
 
-2017-10-02
+2017-10-12
 ----------
 
+* Upgrade Docker version to 17.09 and Docker Compose to 1.16.1 in dind-aws image
 * Upgrade Java version: 8u144
 * Upgrade PHP version: 5.6.31, 7.1.9
 * Upgrade PHP Redis: 3.1.4